backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy[asyncio]==2.0.36
aiomysql==0.2.0
pymysql==1.1.2
alembic==1.14.1
pydantic[email]==2.5.0
pydantic-settings==2.7.1
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==46.0.5
bcrypt==4.2.1
groq==1.1.2
PyJWT==2.12.1
fastapi-mail==1.4.1
